backend/requirements.txt
# Playarr Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.13.0
celery[redis]>=5.3.6
redis>=5.0.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.2
beautifulsoup4>=4.12.2
musicbrainzngs>=0.7.1
yt-dlp>=2024.1.0
python-multipart>=0.0.6
aiofiles>=23.2.1
fake-useragent>=1.4.0
natsort>=8.4.0
Pillow>=10.1.0
python-dotenv>=1.0.0
